Day 17, more faster
As I expected (see last commit message) one way to improve the run time is to
avvoid using Enqueue/Dequeue when not necessary. This new iteration checks
the cache BEFORE enqueuing to avoid enqueuing/dequeuing paths that we already
know are incorrect. This saves time because maintaining the min-heap property
is rather expensive so an Enqueue+Dequeue combo adds a lot of reshuffling of
the heap.

This code now looks properly terrible but it runs 40% faster than the last
iteration.