Day 17 final optimisaiton
Reduce number of useless steps and the complexity of the caching by directly
enqueuing all possible destinations in the one direction. So now instead of
checking if we can carry on or if we want to turn we simply add always turn and
for each direction (left/right) we enqueue all possible jumps, while collecting
the heatloss for that step.

This ends up being about 60% faster than the previous iteration.