Day 21 2023
MEH MEH MEH Really unsatisfied with this... Impossible to solve without
looking at the input. I figured that there was some free path to the edges
but it really didn't help me. I had to look at the solution thread to realise
that this meant the function is a quadratic with a period of the grid size.

The intuition is that you can reach every point in the grid in a number of
steps equal to the grid size (even the corner since we start from the center).
So since any cell is reachable in at most |grid size| steps the reachable steps
grow like a quadratic (proportional to the area). You can verify that by
computing the first few iterations.

Since that works (the first 3 points are enough to predict the fourth), we can
then just start from target % |grid size|, compute the first 3 iterations, feed
this into a 2nd order Lagrange interpolation and we have our solution.

Again ... MEH! Not super satisfied with the problem or how I solved it